+++ CHANGES.md
@@ -5,6 +5,7 @@
 * Use single non-generic Decoder/Encoder instead of sets of type-specific Decoders/Encoders.
 * Decoders/Encoders are now more flexible, having access to the Response/RequestTemplate respectively.
 * Added Feign.Builder to simplify client customizations without using Dagger.
+* Gson type adapters can be registered as Dagger set bindings.
 
 ### Version 4.4.1
 * Fix NullPointerException on calling equals and hashCode.